feat(stdlib): add an internal hack library for manipulating the table stream
Added an internal library for quick ways to manipulate the table stream
to debug transformations. One hack is `flatten()` which will copy the
data and flatten each table into a single buffer. The other is `order()`
which will read the tables and ensure they are returned in sorted group
key order.