feat(execute/function): create an alternative for constructing new transformations
This code tries to simplify creating new transformations by using
reflection to read arguments rather than custom code that looks very
similar. This method should work for a majority of transformations where
creating the operation spec is a simple mapping of types to their Go
equivalents and potentially a bit more.

The idea is that you create a single struct called the
`TransformationSpec`. This is a `ProcedureSpec` that knows how to create
a transformation from itself. The `TransformationSpec` can define and
customize its parameters using struct tags labeled as `flux` or it can
use the defaults as described in the `ReadArgs` documentation. That can
make implementing some transformations more simple. As an example,
`max()` could be implemented by doing:

    type MaxProcedureSpec struct {
        Tables *functions.TableObject
        Column string
    }

    func (s *MaxProcedureSpec) CreateTransformation(id execute.DatasetID, a execute.Administration) (execute.Transformation, execute.Dataset, error) {
        // construct the transformation using the arguments
    }

This work is experimental and incomplete. Please be sure to thoroughly
test the transformation with different arguments when testing. The
intention is to create a foundation that makes creating transformations
easier, less code intense, and easier to document and teach.