replaced BytecodeProgram.Start with the first opcode
The synthesis step now generates a single opcode: program start. It contains
the side effects as an argument. Can now work backwards (and down) to compute
side effects using opcodes that have been synthesized.

Table find is also generating an opcode and executing it. This is synthesized
during evaluation and evaluated immediately. Another near term goal is to
synthesize this in the first synthesis pass and execute as part of the top
level evaluation. Not to recurse into the bytecode execution.