removed function call from interpreter and addes SCOPE_SET
The interpreter was evaluating function calls in addition to the bytecode
execution. Removed this, which causes additional pararmeter analysis to get
moved to the bytecode exec loop.

Implemented the scope set. At interpreter time we are putting placeholders in
the scope so we can report failed lookups at synthesis time.

Can run basic query, a simple assignment, and the tableFind example