user defined function calling
Catching user defined functions in the call and altering IP appropriately.
Requires noting user-funcs during synthesis and generating code for them. But
we need some hacks to exclude the arguments to functions which will become
dynamic functions. Also skipping anything not in the main package so we don't
try and synthesize funcs in universe, too much fail there with unimplemented
things.