feat(execute): add additional messages to the executor
This adds additional messages to the executor and also adds additional
methods on the message itself. The message now contains `Ack`, `Dup`,
and `SetTags`.

The `Ack` method is used to acknowledge that the message has been
received and processed. This is intended to be called by low level code
to mark that the message was consumed. If the message is passed to
another transport, then the `Ack` method doesn't get called.

The `Dup` method gets called to duplicate a message without consuming
the current message. It allows a message to be duplicated and sent to
two different locations. This is primarily used for multiple successors
for a dataset.

The `SetTags` method is primarily used for opentracing. It sets labels
on the current span so that tracing gets documented with appropriate
labels for each message.