chore: remove very slow benchmarks
This removes benchmarks that are very slow and less useful for measuring
the efficiency of code.

Generally, benchmarks are useful if you can find hotspots in the code
and the benchmark runs many times to average out how long the section of
code takes. There were some benchmarks that could have helped for
profiling, but they ran few times because each iteration of the
benchmark took so long.

This also caused the the `test-bench` job for CI to take a very long
time and with very little value for the time it took to run. This job
just verified that the benchmarks still worked with changes, but they
didn't measure changes in the benchmarks.

This removes the very long benchmarks that had little value in order to
speed up the CI build.