demo: making builds reproducible again by ordering docs.json
docs.json contains a list of packages, and list of members in each package.
These are getting written in to docs.json in an unpredictable order (file
system determinted). This is causing libflux.a to be non-reproducible. IE,
there can be no change to the source code, but libflux.a comes out differently
on subsequent build runs.

This change shows one way to make the build reproducible, by sorting the top
level package list when it's constructed, and using serde serialize_with on the
member hashmap.

These methods are inconsistent. To make them consistent one could either switch
the members field from a hash-map to a bst-map, or or one could use
serialize_with on the top level packages variable.

To verify that builds are reproducible use shasum on libflux.a and/or docs.json
in the target directory.