wip: experiment with a naive implementation of flow based types.
Basically I want `if exists r.a then r.a else 0` to work and allow the
exists condition to be evaluated at compile time so that each branch can
have a different type and so that types of the test do not matter at
runtime. This will make the exists operator much more useful because it
won't add a constraint to the type that it must always exist.