feat(docs): adds linter for Flux source docs
The linter can be run via `cargo run --bin flux-lint-docs <file>` and will
report any errors found in the file.

There is also a new list of package in the stdlib that are exempt from
linting errors in their docs. Currently this list is all packages in the
stdlib. The plan is that once a package has been updated to pass the
linter it will be removed from the list. Once all packages are passing
the list we can remove support for exceptions entirely.