feat(stdlib): add an experimental generate package
The `experimental/generate` package can be used as a way to generate
data quickly so that it can be written to an output and then queried
later. It can also be used directly, but it isn't very performant so
that could be a problem.

The main function of the package, `with`, allows a user to specify
cardinalities for keys and then use a function to generate the key. It
can then generate the rows for a table.

An experimental random number generator package has also been added to
aid in the creation of data. The random number generator can be seeded
by the source or the default one seeded by the engine can be used.